kokoro-tts/requirements.txt
kokoro-onnx
soundfile
numpy
flask
gunicorn
google-cloud-storage
